# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Library API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Books": {
      "get": {
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Books"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Authors": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "authorsHasBooks": {
            "$ref": "#/components/schemas/AuthorsHasBooks"
          }
        },
        "additionalProperties": false
      },
      "AuthorsHasBooks": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "booksIsbn": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "authores": {
            "$ref": "#/components/schemas/Authors"
          },
          "booksIsbnNavigation": {
            "$ref": "#/components/schemas/Books"
          }
        },
        "additionalProperties": false
      },
      "Books": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "editorialId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "sinopsis": {
            "type": "string",
            "nullable": true
          },
          "pages": {
            "type": "string",
            "nullable": true
          },
          "editorial": {
            "$ref": "#/components/schemas/Editorials"
          },
          "authorsHasBooks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorsHasBooks"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Editorials": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "editorialName": {
            "type": "string",
            "nullable": true
          },
          "headquarters": {
            "type": "string",
            "nullable": true
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Books"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}